Create SQANTI3_filter.R
SQANTI3 filter and Machine learning approach. In this case, certain improvements to increase processing speed were included.
Plus, a bug was fixed in terms of high correlation features. Previously, the code, retained and removed all variables with common string letters, now, it really retains and remove the variables with correlation >0.9.