Creating SQANTI3_filter.R
This script has the modifications made by Francisco Pardo Palacios and Leandro Balzano-Nogueira, to make it much faster as well as to add more variables.
Also, a bug was fixed in which now only the variables with correlation larger than 0.9 will be removed. Before all variables with the similar string name were removed.